Revert "[front] Fixed table upsert (#7538)"
This reverts commit 3a9ab0f8dbe26a9def885e3d4358a0e3462a6def.
