+++ front/lib/api/tables.ts
@@ -9,10 +9,10 @@ import type {
 import {
   CoreAPI,
   Err,
+  getSanitizedHeaders,
   getSmallWhitelistedModel,
   guessDelimiter,
   Ok,
-  slugify,
 } from "@dust-tt/types";
 import { parse } from "csv-parse";
 import * as t from "io-ts";
@@ -299,31 +299,6 @@ export async function upsertTableFromCsv({
   return tableRes;
 }
 
-function getSanitizedHeaders(rawHeaders: string[]) {
-  return rawHeaders.reduce<string[]>((acc, curr) => {
-    const slugifiedName = slugify(curr);
-
-    if (!acc.includes(slugifiedName) || !slugifiedName.length) {
-      acc.push(slugifiedName);
-    } else {
-      let conflictResolved = false;
-      for (let i = 2; i < 64; i++) {
-        if (!acc.includes(slugify(`${slugifiedName}_${i}`))) {
-          acc.push(slugify(`${slugifiedName}_${i}`));
-          conflictResolved = true;
-          break;
-        }
-      }
-
-      if (!conflictResolved) {
-        // Ignore this header, push empty value
-        acc.push("");
-      }
-    }
-    return acc;
-  }, []);
-}
-
 export async function rowsFromCsv({
   auth,
   csv,
@@ -374,13 +349,11 @@ export async function rowsFromCsv({
     if (i++ >= rowIndex) {
       const record = anyRecord as string[];
       for (const [i, h] of header.entries()) {
-        if (h) {
-          const col = record[i] || "";
-          if (!valuesByCol[h]) {
-            valuesByCol[h] = [col];
-          } else {
-            (valuesByCol[h] as string[]).push(col);
-          }
+        const col = record[i] || "";
+        if (!valuesByCol[h]) {
+          valuesByCol[h] = [col];
+        } else {
+          (valuesByCol[h] as string[]).push(col);
         }
       }
     }
@@ -478,13 +451,9 @@ export async function rowsFromCsv({
   for (let i = 0; i < nbRows; i++) {
     const record = header.reduce(
       (acc, h) => {
-        if (h) {
-          const parsedValues = parsedValuesByCol[h];
-          acc[h] =
-            parsedValues && parsedValues[i] !== undefined
-              ? parsedValues[i]
-              : "";
-        }
+        const parsedValues = parsedValuesByCol[h];
+        acc[h] =
+          parsedValues && parsedValues[i] !== undefined ? parsedValues[i] : "";
         return acc;
       },
       {} as Record<string, CoreAPIRowValue>
